refactor(optimizer): keep closure machinery private to the pass
active_width_closure.h/.cc existed to share readiness and closure
bookkeeping with a research tool outside this repo. That tool no
longer lives here, and nothing else in-tree used it, so fold
is_expanding, UndoStep, SearchFrontier, undo_all,
find_ready_non_expanding, and run_closure into the anonymous namespace
of active_width_schedule_pass.cc, drop the unused apply_op wrapper,
and delete the two files (and their CMake registrations in both
src/clifft and src/wasm). classify_and_apply, is_expanding_effect,
pauli_body, and dense_work_contribution stay in active_width_analysis.h
where the pass already draws them from.

The closure theorem's long sketch collapses to a short invariants
comment (ready, expanding, closure, confluence) pointing to
docs/theory/active-width.md for the argument.
