+++ src/clifft/CMakeLists.txt
@@ -9,7 +9,6 @@ add_library(clifft_core STATIC
     frontend/frontend.cc
     optimizer/hir_pass_manager.cc
     optimizer/active_width_analysis.cc
-    optimizer/active_width_closure.cc
     optimizer/drop_non_unitary_pass.cc
     optimizer/commutation.cc
     optimizer/peephole.cc
